% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{ci.DDL}
\alias{ci.DDL}
\title{Computing confidence intervals}
\usage{
\method{ci}{DDL}(x, alpha = 0.05, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x}{An object of class 'DDL'}

\item{alpha}{alpha Level of significance to construct confidence interval}

\item{alternative}{indicates the alternative hypothesis to construct confidence interval and must be one of "two.sided" (default), "less", or "greater".}
}
\description{
'ci' method for class 'DDL'
}
