% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.summary.DDL}
\alias{print.summary.DDL}
\title{Summarizing DDL}
\usage{
\method{print}{summary.DDL}(x, ...)
}
\arguments{
\item{x}{An object of class 'summary.DDL'}

\item{...}{Ignored}
}
\description{
'summary' method for class 'DDL'
}
