% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{summary.DDL}
\alias{summary.DDL}
\title{Summarizing DDL}
\usage{
\method{summary}{DDL}(object, ...)
}
\arguments{
\item{object}{An object of class 'DDL'}

\item{...}{Ignored}
}
\value{
The function 'summary.DDL' returns a list of summary statistics of DDL given 'DDL'
}
\description{
'summary' method for class 'DDL'
}
